feat: use SDK for registration
* refactored @meinte -code to useRegistrationProcess  hook
* server side generation of more of registration page
