+++ src/app/user/events/[id]/registration/RegistrationComplete.tsx
@@ -1,7 +1,8 @@
+'use client';
+
 import createTranslation from 'next-translate/createTranslation';
 
 import Button from '@/components/ui/Button';
-import Heading from '@/components/ui/Heading';
 
 export type RegistrationCompleteProps = {
   onSubmit: () => void;
@@ -10,9 +11,8 @@ const RegistrationComplete = ({ onSubmit }: RegistrationCompleteProps) => {
   const { t } = createTranslation();
   return (
     <>
-      <Heading>{t('user:registration.complete.title')}</Heading>
       <p>{t('user:registration.complete.description')}</p>
-      <Button onClick={() => onSubmit()}>{t('common:buttons.continue')}</Button>
+      <Button onClick={() => onSubmit()}>{t('common:buttons.homepage')}</Button>
     </>
   );
 };
