+++ src/components/forms/ProductSelection.tsx
@@ -1,46 +1,46 @@
-import { RegistrationProduct } from '@/types/RegistrationProduct';
+import { ProductDto } from '@losol/eventuras';
 
 import Checkbox, { CheckBoxDescription, CheckBoxLabel } from './Checkbox';
 export type ProductSelectionProps = {
-  products: RegistrationProduct[];
+  products: ProductDto[];
   register: any;
 };
 const ProductSelection = ({ products, register }: ProductSelectionProps) => {
   return (
     <>
-      {products.map((product: RegistrationProduct) => {
-        if (product.isBooleanSelection) {
+      {products.map((product: ProductDto) => {
+        if (!product.enableQuantity) {
           return (
             <Checkbox
-              key={product.id}
-              id={product.id}
-              title={product.title}
+              key={product.productId}
+              id={product.productId}
+              title={product.name}
               description={product.description}
-              {...register(`products.${product.id}`)}
-              disabled={product.mandatory}
-              defaultChecked={product.mandatory}
+              {...register(`products.${product.productId}`)}
+              disabled={product.isMandatory}
+              defaultChecked={product.isMandatory}
             >
-              <CheckBoxLabel>{product.title}</CheckBoxLabel>
+              <CheckBoxLabel>{product.name}</CheckBoxLabel>
               <CheckBoxDescription>{product.description}</CheckBoxDescription>
             </Checkbox>
           );
         }
         return (
-          <div key={product.id}>
+          <div key={product.productId}>
             <input
               type="number"
               className="w-16 mb-3 appearance-none border rounded py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
               defaultValue={product.minimumQuantity}
               min={product.minimumQuantity}
-              value={product.id}
-              {...register(`products.${product.id}`, {
+              value={product.productId}
+              {...register(`products.${product.productId}`, {
                 min: {
                   value: product.minimumQuantity,
                   message: `Minimum is ${product.minimumQuantity}`,
                 },
               })}
             />
-            <label htmlFor={product.id}>{product.title}</label>
+            <label htmlFor={product.productId?.toString()}>{product.name}</label>
             <p>{product.description}</p>
           </div>
         );
